refactor: extract ac view model
